separate modules for consensus snapshot chunks
